+++ main.cpp
@@ -12,6 +12,7 @@
 #include "PSphere.h"
 #include "OgreConfigFile.h"
 #include "GeneratorFrameListener.h"
+#include "initOgre.h"
 #include <OgreMeshSerializer.h>
 
 #if OGRE_PLATFORM == OGRE_PLATFORM_WIN32
@@ -23,121 +24,121 @@
 #include <sys/stat.h>
 #endif
 
- class initOgre
- {
- public:
-	Ogre::Root              *Root;
-	Ogre::SceneManager      *Scene;
-	Ogre::SceneNode         *RootSceneNode;
-	Ogre::RenderWindow      *Window;
-	Ogre::Camera            *Camera;
-	GeneratorFrameListener  *FrameListener;
-	Ogre::OverlaySystem     *OverlaySystem;
+//  class initOgre
+//  {
+//  public:
+// 	Ogre::Root              *Root;
+// 	Ogre::SceneManager      *Scene;
+// 	Ogre::SceneNode         *RootSceneNode;
+// 	Ogre::RenderWindow      *Window;
+// 	Ogre::Camera            *Camera;
+// 	GeneratorFrameListener  *FrameListener;
+// 	Ogre::OverlaySystem     *OverlaySystem;
 
-	initOgre()
-	{
-		FrameListener = 0;
-		Root = 0;
-		OverlaySystem = 0;
+// 	initOgre()
+// 	{
+// 		FrameListener = 0;
+// 		Root = 0;
+// 		OverlaySystem = 0;
 
-	}
-	int start();
-	void CreateFrameListener();
+// 	}
+// 	int start();
+// 	void CreateFrameListener();
 
-};
+// };
 
-int initOgre::start()
-{
+// int initOgre::start()
+// {
 
-	Root = new Ogre::Root("", "", "Generator.LOG");
-	OverlaySystem = new Ogre::OverlaySystem();
-
-
-	Ogre::String PluginName;
-#if OGRE_PLATFORM == OGRE_PLATFORM_WIN32
-	PluginName.append("RenderSystem_Direct3D9");
-#else
-	/* Use Posix-function stat to check existence of a file.
-	 * This is needed because Ogre makes zero effort to find its own plugins. */
-	struct stat statBuf;
-	Ogre::String PluginFile[3];
-
-	PluginFile[0].append("/usr/lib/OGRE/RenderSystem_GL.so");
-	PluginFile[1].append("/usr/lib/local/OGRE/RenderSystem_GL.so");
-	// Ubuntu
-#ifdef __x86_64__
-	PluginFile[2].append("/usr/lib/x86_64-linux-gnu/OGRE-");
-#elif __i386__
-	PluginFile[2].append("/usr/lib/i386-linux-gnu/OGRE-");
-#endif
-	PluginFile[2] += Ogre::StringConverter::toString(OGRE_VERSION_MAJOR) + ".";
-	PluginFile[2] += Ogre::StringConverter::toString(OGRE_VERSION_MINOR) + ".";
-	PluginFile[2] += Ogre::StringConverter::toString(OGRE_VERSION_PATCH);
-	PluginFile[2] += "/RenderSystem_GL.so";
-
-	int i;
-	for(i=0; i < 3; i++)
-	{
-		if( stat(PluginFile[i].c_str(), &statBuf) == 0 )
-		{
-			PluginFile[i].resize(PluginFile[i].size()-3);	// strip ".so"
-			PluginName.assign( PluginFile[i] );
-			break;											// Exit loop if file exists
-		}
-	}
-	if(PluginName == "")
-		PluginName.append("RenderSystem_GL");					// When all else fails...
-#endif
-	if(OGRE_DEBUG_MODE)
-		PluginName.append("_d");    // RenderSystem_GL_d if using Ogre debug mode
-
-	// Loads renderer plugin
-	Root->loadPlugin(PluginName);
+// 	Root = new Ogre::Root("", "", "Generator.LOG");
+// 	OverlaySystem = new Ogre::OverlaySystem();
 
 
-	// Check renderer availability
-	const Ogre::RenderSystemList& RenderSystemList = Root->getAvailableRenderers();
-	if( RenderSystemList.size() == 0 )
-	{
-		Ogre::LogManager::getSingleton().logMessage("Rendersystems not found.");
-		return 1;
-	}
-	Ogre::RenderSystem *RenderSystem = RenderSystemList[0];
-	Root->setRenderSystem(RenderSystem);
+// 	Ogre::String PluginName;
+// #if OGRE_PLATFORM == OGRE_PLATFORM_WIN32
+// 	PluginName.append("RenderSystem_Direct3D9");
+// #else
+// 	/* Use Posix-function stat to check existence of a file.
+// 	 * This is needed because Ogre makes zero effort to find its own plugins. */
+// 	struct stat statBuf;
+// 	Ogre::String PluginFile[3];
+
+// 	PluginFile[0].append("/usr/lib/OGRE/RenderSystem_GL.so");
+// 	PluginFile[1].append("/usr/lib/local/OGRE/RenderSystem_GL.so");
+// 	// Ubuntu
+// #ifdef __x86_64__
+// 	PluginFile[2].append("/usr/lib/x86_64-linux-gnu/OGRE-");
+// #elif __i386__
+// 	PluginFile[2].append("/usr/lib/i386-linux-gnu/OGRE-");
+// #endif
+// 	PluginFile[2] += Ogre::StringConverter::toString(OGRE_VERSION_MAJOR) + ".";
+// 	PluginFile[2] += Ogre::StringConverter::toString(OGRE_VERSION_MINOR) + ".";
+// 	PluginFile[2] += Ogre::StringConverter::toString(OGRE_VERSION_PATCH);
+// 	PluginFile[2] += "/RenderSystem_GL.so";
+
+// 	int i;
+// 	for(i=0; i < 3; i++)
+// 	{
+// 		if( stat(PluginFile[i].c_str(), &statBuf) == 0 )
+// 		{
+// 			PluginFile[i].resize(PluginFile[i].size()-3);	// strip ".so"
+// 			PluginName.assign( PluginFile[i] );
+// 			break;											// Exit loop if file exists
+// 		}
+// 	}
+// 	if(PluginName == "")
+// 		PluginName.append("RenderSystem_GL");					// When all else fails...
+// #endif
+// 	if(OGRE_DEBUG_MODE)
+// 		PluginName.append("_d");    // RenderSystem_GL_d if using Ogre debug mode
+
+// 	// Loads renderer plugin
+// 	Root->loadPlugin(PluginName);
 
 
-	// Do not create window automatically
-	Root->initialise(false);
+// 	// Check renderer availability
+// 	const Ogre::RenderSystemList& RenderSystemList = Root->getAvailableRenderers();
+// 	if( RenderSystemList.size() == 0 )
+// 	{
+// 		Ogre::LogManager::getSingleton().logMessage("Rendersystems not found.");
+// 		return 1;
+// 	}
+// 	Ogre::RenderSystem *RenderSystem = RenderSystemList[0];
+// 	Root->setRenderSystem(RenderSystem);
 
-	Ogre::NameValuePairList renderParams;
 
-	renderParams["FSAA"] = "0";
-	renderParams["vsync"] = "true"; // Waits for vertical sync
+// 	// Do not create window automatically
+// 	Root->initialise(false);
 
-	// Actual call to create window, bool value is fullscreen flag
-	Window = Root->createRenderWindow("My little planet", 800, 600, false, &renderParams);
+// 	Ogre::NameValuePairList renderParams;
+
+// 	renderParams["FSAA"] = "0";
+// 	renderParams["vsync"] = "true"; // Waits for vertical sync
+
+// 	// Actual call to create window, bool value is fullscreen flag
+// 	Window = Root->createRenderWindow("My little planet", 800, 600, false, &renderParams);
 
 
-	// Start creating scene
-	Scene = Root->createSceneManager(Ogre::ST_GENERIC);
-	RootSceneNode = Scene->getRootSceneNode();
+// 	// Start creating scene
+// 	Scene = Root->createSceneManager(Ogre::ST_GENERIC);
+// 	RootSceneNode = Scene->getRootSceneNode();
 
-	if(OverlaySystem)
-		Scene->addRenderQueueListener(OverlaySystem);
+// 	if(OverlaySystem)
+// 		Scene->addRenderQueueListener(OverlaySystem);
 
-	return 0;
-}
-void initOgre::CreateFrameListener()
-{
+// 	return 0;
+// }
+// void initOgre::CreateFrameListener()
+// {
 
-#if OGRE_PLATFORM == OGRE_PLATFORM_APPLE_IOS
-	FrameListener= new GeneratorFrameListener(Window, Camera, true, true, true);
-#else
-	FrameListener= new GeneratorFrameListener(Window, Camera,RootSceneNode);
-#endif
-	FrameListener->showDebugOverlay(true);
-	Root->addFrameListener(FrameListener);
-}
+// #if OGRE_PLATFORM == OGRE_PLATFORM_APPLE_IOS
+// 	FrameListener= new GeneratorFrameListener(Window, Camera, true, true, true);
+// #else
+// 	FrameListener= new GeneratorFrameListener(Window, Camera,RootSceneNode);
+// #endif
+// 	FrameListener->showDebugOverlay(true);
+// 	Root->addFrameListener(FrameListener);
+// }
 
 
 
@@ -185,10 +186,13 @@ int main(int argc, char *argv[])
 		PSphere mySphere;
 		mySphere.create(15.0f, 0.6f, 100);
 		mySphere.loadToBuffers("CustomMesh", "sphereTex");
+
+		//Export the shape in a mesh file before destroying it
 		Ogre::MeshPtr mesh;
 		mesh = mySphere.getMesh();
 		Ogre::MeshSerializer ser;
-		ser.exportMesh(mesh.getPointer(), "C:\\Users\\giova\\Documents\\PlanetGenerator\\planet.mesh",  Ogre::MeshSerializer::ENDIAN_NATIVE);
+		//ser.exportMesh(mesh.getPointer(), "C:\\Users\\giova\\Documents\\PlanetGenerator\\planet.mesh",  Ogre::MeshSerializer::ENDIAN_NATIVE);
+
 		mySphere.destroy();
 
 		Ogre::Entity *entity1 = myOgre.Scene->createEntity("CustomEntity", "CustomMesh");
